requirements.txt
tensordict-nightly
torchrl-nightly
tqdm
hydra-core>=1.1
hydra-submitit-launcher
wandb
pettingzoo
jsbsim==1.2.1
scipy
matplotlib